+++ bootique-jdbc-tomcat-instrumented/src/main/java/io/bootique/jdbc/instrumented/tomcat/TomcatInstrumentedJdbcModuleProvider.java
@@ -3,9 +3,9 @@ package io.bootique.jdbc.instrumented.tomcat;
 import com.google.inject.Module;
 import io.bootique.BQModule;
 import io.bootique.BQModuleProvider;
-import io.bootique.jdbc.instrumented.InstrumentedJdbcModuleProvider;
 import io.bootique.jdbc.tomcat.TomcatJdbcModuleProvider;
 import io.bootique.metrics.MetricsModuleProvider;
+import io.bootique.metrics.health.HealthCheckModuleProvider;
 
 import java.util.Collection;
 
@@ -29,8 +29,8 @@ public class TomcatInstrumentedJdbcModuleProvider implements BQModuleProvider {
     public Collection<BQModuleProvider> dependencies() {
         return asList(
                 new TomcatJdbcModuleProvider(),
-                new InstrumentedJdbcModuleProvider(),
-                new MetricsModuleProvider()
+                new MetricsModuleProvider(),
+                new HealthCheckModuleProvider()
         );
     }
 }