+++ bootique-jdbc-tomcat-instrumented/src/main/java/io/bootique/jdbc/instrumented/tomcat/healthcheck/DataSourceHealthCheck.java
@@ -1,4 +1,4 @@
-package io.bootique.jdbc.instrumented.healthcheck;
+package io.bootique.jdbc.instrumented.tomcat.healthcheck;
 
 import io.bootique.metrics.health.HealthCheck;
 import io.bootique.metrics.health.HealthCheckOutcome;