+++ bootique-jdbc-tomcat-instrumented/src/test/java/io/bootique/jdbc/instrumented/tomcat/TomcatInstrumentedJdbcModuleProviderTest.java
@@ -2,7 +2,6 @@ package io.bootique.jdbc.instrumented.tomcat;
 
 import io.bootique.BQRuntime;
 import io.bootique.jdbc.JdbcModule;
-import io.bootique.jdbc.instrumented.InstrumentedJdbcModule;
 import io.bootique.jdbc.tomcat.TomcatJdbcModule;
 import io.bootique.metrics.MetricsModule;
 import io.bootique.test.junit.BQModuleProviderChecker;
@@ -27,7 +26,6 @@ public class TomcatInstrumentedJdbcModuleProviderTest {
         BQRuntimeChecker.testModulesLoaded(bqRuntime,
                 JdbcModule.class,
                 TomcatJdbcModule.class,
-                InstrumentedJdbcModule.class,
                 MetricsModule.class
         );
     }